+++ src/lib.rs
@@ -22,8 +22,8 @@ use wasm_bindgen::prelude::*;
 #[cfg(feature = "fetch")]
 /// Builds the default headers for the client.
 fn build_headers() -> reqwest::header::HeaderMap {
-    use reqwest::header;
     use header::{HeaderMap, HeaderValue};
+    use reqwest::header;
     let mut headers = HeaderMap::new();
     headers.insert(
         header::USER_AGENT,
@@ -45,8 +45,8 @@ fn build_headers() -> reqwest::header::HeaderMap {
 }
 
 #[cfg(feature = "wasm_parser")]
-#[derive(tsify::Tsify, serde::Deserialize, serde::Serialize)]
-#[tsify(into_wasm_abi)]
+#[derive(serde::Deserialize, serde::Serialize)]
+#[cfg_attr(feature = "wasm_parser", derive(tsify::Tsify), tsify(into_wasm_abi))]
 /// Wrapper of the search results vector.
 pub struct Search(Vec<search::SearchResult>);
 
@@ -54,6 +54,7 @@ pub struct Search(Vec<search::SearchResult>);
 #[wasm_bindgen]
 /// Parses the HTML body and returns the search results.
 pub fn parse_search_results(webpage_body: String) -> Search {
+    console_error_panic_hook::set_once();
     let search_results = ProductSearch::parse(webpage_body);
     Search(search_results)
 }
@@ -62,5 +63,14 @@ pub fn parse_search_results(webpage_body: String) -> Search {
 #[wasm_bindgen]
 /// Parses the HTML body and returns the product details.
 pub fn parse_product_details(webpage_body: String) -> Result<ProductDetails, JsError> {
+    console_error_panic_hook::set_once();
     ProductDetails::parse(webpage_body).map_err(|e| JsError::new(&e.to_string()))
+}
+
+#[cfg(feature = "wasm_parser")]
+#[wasm_bindgen]
+pub fn build_search_url(query: String, params: &search::SearchParams) -> Result<String, JsError> {
+    ProductSearch::build_request_url(query, params)
+        .map(|url| url.to_string())
+        .map_err(|e| JsError::new(&e.to_string()))
 }