+++ src/product_details/seller.rs
@@ -1,5 +1,5 @@
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "wasm_parser", derive(tsify::Tsify), tsify(into_wasm_abi))]
+#[cfg_attr(feature = "wasm_parser", derive(tsify::Tsify))]
 #[derive(Debug)]
 /// Information about the seller of a Product.
 pub struct Seller {